backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
fastapi==0.135.2
filelock==3.25.2
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
openai-whisper==20250625
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
reportlab==4.4.10
requests==2.33.0
setuptools==81.0.0
SQLAlchemy==2.0.48
starlette==1.0.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
